Campaign metadata.

Campaign metadata for all campaigns run for the Customer Journey study. This dataset gives the length of time for which a campaign runs. So, any coupons received as part of a campaign are valid within the dates contained in this dataset.

campaign_descriptions

Format

A data frame with 27 rows and 4 variables

campaign_id: Uniquely identifies each campaign; Ranges 1-27

campaign_type: Type of campaign (Type A, Type B, Type C)

start_date: Start date of campaign

end_date: End date of campaign

Source

84.51°, Customer Journey study, http://www.8451.com/area51/

Value

campaign_descriptions

a tibble

Examples

# \donttest{
# full data set
campaign_descriptions
#> # A tibble: 27 × 4
#>    campaign_id campaign_type start_date end_date  
#>    <chr>       <ord>         <date>     <date>    
#>  1 1           Type B        2017-03-03 2017-04-09
#>  2 2           Type B        2017-03-08 2017-04-09
#>  3 3           Type C        2017-03-13 2017-05-08
#>  4 4           Type B        2017-03-29 2017-04-30
#>  5 5           Type B        2017-04-03 2017-05-07
#>  6 6           Type C        2017-04-19 2017-05-21
#>  7 7           Type B        2017-04-24 2017-05-28
#>  8 8           Type A        2017-05-08 2017-06-25
#>  9 9           Type B        2017-05-31 2017-07-02
#> 10 10          Type B        2017-06-28 2017-07-30
#> # ℹ 17 more rows

# Join product campaign metadata to campaign_table dataset
require("dplyr")
#> Loading required package: dplyr
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
campaigns %>%
  left_join(campaign_descriptions, "campaign_id")
#> # A tibble: 6,589 × 5
#>    campaign_id household_id campaign_type start_date end_date  
#>    <chr>       <chr>        <ord>         <date>     <date>    
#>  1 1           105          Type B        2017-03-03 2017-04-09
#>  2 1           1238         Type B        2017-03-03 2017-04-09
#>  3 1           1258         Type B        2017-03-03 2017-04-09
#>  4 1           1483         Type B        2017-03-03 2017-04-09
#>  5 1           2200         Type B        2017-03-03 2017-04-09
#>  6 1           293          Type B        2017-03-03 2017-04-09
#>  7 1           529          Type B        2017-03-03 2017-04-09
#>  8 1           536          Type B        2017-03-03 2017-04-09
#>  9 1           568          Type B        2017-03-03 2017-04-09
#> 10 1           630          Type B        2017-03-03 2017-04-09
#> # ℹ 6,579 more rows
# }