Coupon metadata.

Coupon metadata for all coupons used in campaigns advertised to households participating in the Customer Journey study.

coupons

Format

A data frame with 116,204 rows and 3 variables

coupon_upc

Uniquely identifies each coupon (unique to household and campaign)

product_id

Uniquely identifies each product

campaign_id

Uniquely identifies each campaign

Source

84.51°, Customer Journey study, http://www.8451.com/area51/

Examples

# if data hasn't been imported yet
get_data(which = "all", verbose = FALSE)

coupon

#> Error in eval(expr, envir, enclos): object 'coupon' not found

# Join product metadata to coupon dataset
require("dplyr")
coupons %>%
  left_join(products, "product_id")

#> # A tibble: 116,204 x 9
#>    coupon_upc product_id campaign_id manufacturer_id department brand
#>    <chr>      <chr>      <chr>       <chr>           <chr>      <fct>
#>  1 100000852… 9676830    26          1722            GROCERY    Nati…
#>  2 100000852… 9676943    26          317             GROCERY    Nati…
#>  3 100000852… 9676944    26          317             GROCERY    Nati…
#>  4 100000852… 9676947    26          1722            GROCERY    Nati…
#>  5 100000852… 9677008    26          1722            GROCERY    Nati…
#>  6 100000852… 9677052    26          317             GROCERY    Nati…
#>  7 100000852… 9677385    26          317             GROCERY    Nati…
#>  8 100000852… 9677479    26          317             GROCERY    Nati…
#>  9 100000852… 9677791    26          317             GROCERY    Nati…
#> 10 100000852… 9677878    26          317             GROCERY    Nati…
#> # … with 116,194 more rows, and 3 more variables: product_category <chr>,
#> #   product_type <chr>, package_size <chr>