Get full Complete Journey transactions data set.

The complete transactions data set for the Complete Journey is too large to be contained within the package. get_transactions() provides an efficient method for downloading the full data set from the source GitHub repository.

get_transactions(verbose = FALSE)

Source

Downloading from https://github.com/bradleyboehmke/completejourney/tree/master/data. Data originated from 84.51°, Customer Journey study, http://www.8451.com/area51/ and were processes for analysis.

Arguments

verbose

Logical indicator whether or not to download silently.

Value

A data frame with 1,469,307 rows and 11 variables. Returns NULL if the download fails (e.g., network timeout, GitHub unavailability) with an informative message about the failure.

See also

Examples

# \donttest{
# requires internet connection
transactions <- get_transactions()
#> Warning: cannot open URL 'https://github.com/bradleyboehmke/completejourney/blob/master/data/transactions.rds?raw=true': HTTP status was '503 Service Unavailable'
#> 
#> Unable to download 'transactions' data from GitHub.
#> This could be due to:
#>   - Network connectivity issues
#>   - GitHub service unavailability
#>   - Repository changes or removal
#> 
#> Original error: cannot open the connection to 'https://github.com/bradleyboehmke/completejourney/blob/master/data/transactions.rds?raw=true'
#> 
#> Please try again later or check your internet connection.
#> Learn more at: https://github.com/bradleyboehmke/completejourney
# }