Promotions data.

Promotions data of the Complete Journey study signifying whether a given product was featured in the weekly mailer or was part of an in-store display (other than regular product placement).

promotions

Format

A data frame with 20,940,529 rows and 5 variables

product_id

Uniquely identifies each product

store_id

Uniquely identifies each store

display_location

Display location (see details for range of values)

mailer_location

Mailer location (see details for range of values)

week

Week of the transaction; Ranges 1-53

Source

84.51°, Customer Journey study, http://www.8451.com/area51/

Display Location Codes

0 - Not on Display

1 - Store Front

2 - Store Rear

3 - Front End Cap

4 - Mid-Aisle End Cap

5 - Rear End Cap

6 - Side-Aisle End Cap

7 - In-Aisle

9 - Secondary Location Display

A - In-Shelf

Mailer Location Codes

0 - Not on ad

A - Interior page feature

C - Interior page line item

D - Front page feature

F - Back page feature

H - Wrap from feature

J - Wrap interior coupon

L - Wrap back feature

P - Interior page coupon

X - Free on interior page

Z - Free on front page, back page or wrap

Examples

# if data hasn't been imported yet
get_data(which = "all", verbose = FALSE)

promotions

#> # A tibble: 20,940,529 x 5
#>    product_id store_id display_location mailer_location  week
#>    <chr>      <chr>    <fct>            <fct>           <int>
#>  1 1000050    316      9                0                   1
#>  2 1000050    337      3                0                   1
#>  3 1000050    441      5                0                   1
#>  4 1000092    292      0                A                   1
#>  5 1000092    293      0                A                   1
#>  6 1000092    295      0                A                   1
#>  7 1000092    298      0                A                   1
#>  8 1000092    299      0                A                   1
#>  9 1000092    304      0                A                   1
#> 10 1000092    306      0                A                   1
#> # … with 20,940,519 more rows

# Join promotions to transactions to analyze
# product promotion/location
require("dplyr")
transactions %>%
  left_join(promotions,
            c("product_id", "store_id", "week"))

#> # A tibble: 1,470,002 x 13
#>    household_id store_id basket_id product_id quantity sales_value retail_disc
#>    <chr>        <chr>    <chr>     <chr>         <dbl>       <dbl>       <dbl>
#>  1 900          330      31198570… 1095275           1        0.5        0    
#>  2 900          330      31198570… 9878513           1        0.99       0.1  
#>  3 1228         406      31198655… 1041453           1        1.43       0.15 
#>  4 906          319      31198705… 1020156           1        1.5        0.290
#>  5 906          319      31198705… 1053875           2        2.78       0.8  
#>  6 906          319      31198705… 1060312           1        5.49       0.5  
#>  7 906          319      31198705… 1075313           1        1.5        0.290
#>  8 1058         381      31198676… 985893            1        1.88       0.21 
#>  9 1058         381      31198676… 988791            1        1.5        1.29 
#> 10 1058         381      31198676… 9297106           1        2.69       0    
#> # … with 1,469,992 more rows, and 6 more variables: coupon_disc <dbl>,
#> #   coupon_match_disc <dbl>, week <int>, transaction_timestamp <dttm>,
#> #   display_location <fct>, mailer_location <fct>