Sampling of the full promotions data set.

A sampling of the promotions data from the Complete Journey study signifying whether a given product was featured in the weekly mailer or was part of an in-store display (other than regular product placement).

promotions_sample

Format

A data frame with 1,948,597 rows and 5 variables

product_id: Uniquely identifies each product

store_id: Uniquely identifies each store

display_location: Display location (see details for range of values)

mailer_location: Mailer location (see details for range of values)

week: Week of the transaction; Ranges 1-53

Source

84.51°, Customer Journey study, http://www.8451.com/area51/

Display Location Codes

0 - Not on Display

1 - Store Front

2 - Store Rear

3 - Front End Cap

4 - Mid-Aisle End Cap

5 - Rear End Cap

6 - Side-Aisle End Cap

7 - In-Aisle

9 - Secondary Location Display

A - In-Shelf

Mailer Location Codes

0 - Not on ad

A - Interior page feature

C - Interior page line item

D - Front page feature

F - Back page feature

H - Wrap from feature

J - Wrap interior coupon

L - Wrap back feature

P - Interior page coupon

X - Free on interior page

Z - Free on front page, back page or wrap

See also

Examples

# sampled promotions data set
promotions_sample

#> # A tibble: 1,948,597 x 5
#>    product_id store_id display_location mailer_location  week
#>    <chr>      <chr>    <fct>            <fct>           <int>
#>  1 1000050    337      3                0                   1
#>  2 1000092    317      0                A                   1
#>  3 1000092    329      0                A                   1
#>  4 1000092    343      0                A                   1
#>  5 1000092    356      0                A                   1
#>  6 1000092    358      0                A                   1
#>  7 1000092    359      0                A                   1
#>  8 1000092    362      0                A                   1
#>  9 1000092    369      0                A                   1
#> 10 1000214    317      6                0                   1
#> # … with 1,948,587 more rows

# Join promotions to transactions to analyze
# product promotion/location
require("dplyr")
transactions %>%
  left_join(promotions_sample,
            c("product_id", "store_id", "week"))

#> Error in eval(lhs, parent, parent): object 'transactions' not found