Sampling of the full Complete Journey transactions.

A sampling of all products purchased by households within the Complete Journey study. Each line found in this table is essentially the same line that would be found on a store receipt. This is only a subsample of the complete data set to keep package size manageable.

transactions_sample

Format

A data frame with 100,000 rows and 11 variables

household_id

Uniquely identifies each household

store_id

Uniquely identifies each store

basket_id

Uniquely identifies a purchase occasion

product_id

Uniquely identifies each product

quantity

Number of the products purchased during the trip

sales_value

Amount of dollars retailer receives from sale

retail_disc

Discount applied due to retailer's loyalty card program

coupon_disc

Discount applied due to manufacturer coupon

coupon_match_disc

Discount applied due to retailer's match of manufacturer coupon

week

Week of the transaction; Ranges 1-53

transaction_timestamp

Date and time of when the transaction occurred

Source

84.51°, Customer Journey study, http://www.8451.com/area51/

See also

Examples

transactions_sample

#> # A tibble: 100,000 x 11
#>    household_id store_id basket_id product_id quantity sales_value retail_disc
#>    <chr>        <chr>    <chr>     <chr>         <dbl>       <dbl>       <dbl>
#>  1 2261         309      31625220… 940996            1        3.86        0.43
#>  2 2131         368      32053127… 873902            1        1.59        0.9 
#>  3 511          316      32445856… 847901            1        1           0.69
#>  4 400          388      31932241… 13094913          2       11.9         2.9 
#>  5 918          340      32074655… 1085604           1        1.29        0   
#>  6 718          324      32614612… 883203            1        2.5         0.49
#>  7 868          323      32074722… 9884484           1        3.49        0   
#>  8 1688         450      34850403… 1028715           1        2           1.79
#>  9 467          31782    31280745… 896613            2        6.55        4.44
#> 10 1947         32004    32744181… 978497            1        3.99        0   
#> # … with 99,990 more rows, and 4 more variables: coupon_disc <dbl>,
#> #   coupon_match_disc <dbl>, week <int>, transaction_timestamp <dttm>