Promotions data.

Promotions data of the Complete Journey study signifying whether a given product was featured in the weekly mailer or was part of an in-store display (other than regular product placement).

promotions

Format

A data frame with 20,940,529 rows and 5 variables

product_id

Uniquely identifies each product

store_id

Uniquely identifies each store

display_location

Display location (see details for range of values)

mailer_location

Mailer location (see details for range of values)

week

Week of the transaction; Ranges 1-53

Source

84.51°, Customer Journey study, http://www.8451.com/area51/

Display Location Codes

0 - Not on Display

1 - Store Front

2 - Store Rear

3 - Front End Cap

4 - Mid-Aisle End Cap

5 - Rear End Cap

6 - Side-Aisle End Cap

7 - In-Aisle

9 - Secondary Location Display

A - In-Shelf

Mailer Location Codes

0 - Not on ad

A - Interior page feature

C - Interior page line item

D - Front page feature

F - Back page feature

H - Wrap from feature

J - Wrap interior coupon

L - Wrap back feature

P - Interior page coupon

X - Free on interior page

Z - Free on front page, back page or wrap

Examples

# NOT RUN {
if (require("dplyr")) {
# if data hasn't been imported yet
import_data(which = "all")

promotions

# Join promotions to transactions to analyze
# product promotion/location
transactions %>%
  left_join(promotions,
            c("product_id", "store_id", "week"))
}
# }