Sampling of the full promotions data set.

A sampling of the promotions data from the Complete Journey study signifying whether a given product was featured in the weekly mailer or was part of an in-store display (other than regular product placement).

promotions_sample

Format

A data frame with 360,535 rows and 5 variables

product_id: Uniquely identifies each product

store_id: Uniquely identifies each store

display_location: Display location (see details for range of values)

mailer_location: Mailer location (see details for range of values)

week: Week of the transaction; Ranges 1-53

Source

84.51°, Customer Journey study, http://www.8451.com/area51/

Value

promotions_sample

a tibble

Display Location Codes

0 - Not on Display

1 - Store Front

2 - Store Rear

3 - Front End Cap

4 - Mid-Aisle End Cap

5 - Rear End Cap

6 - Side-Aisle End Cap

7 - In-Aisle

9 - Secondary Location Display

A - In-Shelf

Mailer Location Codes

0 - Not on ad

A - Interior page feature

C - Interior page line item

D - Front page feature

F - Back page feature

H - Wrap from feature

J - Wrap interior coupon

L - Wrap back feature

P - Interior page coupon

X - Free on interior page

Z - Free on front page, back page or wrap

See also

Examples

# \donttest{
# sampled promotions data set
promotions_sample
#> # A tibble: 360,535 × 5
#>    product_id store_id display_location mailer_location  week
#>    <chr>      <chr>    <fct>            <fct>           <int>
#>  1 1000050    337      3                0                   1
#>  2 1000092    317      0                A                   1
#>  3 1000214    317      6                0                   1
#>  4 1000235    317      0                A                   1
#>  5 1000235    337      0                A                   1
#>  6 1000343    317      9                0                   1
#>  7 1000365    317      0                A                   1
#>  8 1000365    337      0                A                   1
#>  9 100189     317      0                A                   1
#> 10 100189     337      5                A                   1
#> # ℹ 360,525 more rows

# Join promotions to transactions to analyze
# product promotion/location
require("dplyr")
transactions_sample %>%
  left_join(
    promotions_sample,
    c("product_id", "store_id", "week")
  )
#> # A tibble: 75,000 × 13
#>    household_id store_id basket_id   product_id quantity sales_value retail_disc
#>    <chr>        <chr>    <chr>       <chr>         <dbl>       <dbl>       <dbl>
#>  1 2261         309      31625220889 940996            1        3.86        0.43
#>  2 2131         368      32053127496 873902            1        1.59        0.9 
#>  3 511          316      32445856036 847901            1        1           0.69
#>  4 400          388      31932241118 13094913          2       11.9         2.9 
#>  5 918          340      32074655895 1085604           1        1.29        0   
#>  6 718          324      32614612029 883203            1        2.5         0.49
#>  7 868          323      32074722463 9884484           1        3.49        0   
#>  8 1688         450      34850403304 1028715           1        2           1.79
#>  9 467          31782    31280745102 896613            2        6.55        4.44
#> 10 1947         32004    32744181707 978497            1        3.99        0   
#> # ℹ 74,990 more rows
#> # ℹ 6 more variables: coupon_disc <dbl>, coupon_match_disc <dbl>, week <int>,
#> #   transaction_timestamp <dttm>, display_location <fct>, mailer_location <fct>
# }