Complete Journey transactions.

All products purchased by households within the Complete Journey study. Each line found in this table is essentially the same line that would be found on a store receipt. This is only a subsample of the complete data set to keep package size manageable.

transactions

Format

A data frame with 1,469,307 rows and 11 variables

household_id

Uniquely identifies each household

store_id

Uniquely identifies each store

basket_id

Uniquely identifies a purchase occasion

product_id

Uniquely identifies each product

quantity

Number of the products purchased during the trip

sales_value

Amount of dollars retailer receives from sale

retail_disc

Discount applied due to retailer's loyalty card program

coupon_disc

Discount applied due to manufacturer coupon

coupon_match_disc

Discount applied due to retailer's match of manufacturer coupon

week

Week of the transaction; Ranges 1-53

transaction_timestamp

Date and time of when the transaction occurred

Source

84.51, Customer Journey study, http://www.8451.com/area51/

Examples

  head(transactions)

#> # A tibble: 6 x 11
#>   household_id store_id basket_id product_id quantity sales_value retail_disc
#>   <chr>        <chr>    <chr>     <chr>         <dbl>       <dbl>       <dbl>
#> 1 900          330      31198570… 1095275           1        0.5        0    
#> 2 900          330      31198570… 9878513           1        0.99       0.1  
#> 3 1228         406      31198655… 1041453           1        1.43       0.15 
#> 4 906          319      31198705… 1020156           1        1.5        0.290
#> 5 906          319      31198705… 1053875           2        2.78       0.8  
#> 6 906          319      31198705… 1060312           1        5.49       0.5  
#> # … with 4 more variables: coupon_disc <dbl>, coupon_match_disc <dbl>,
#> #   week <int>, transaction_timestamp <dttm>